improve readability
for me it was not clear at first glance what it does. should be clearer now.
And i stumbled upon the problem that it has to be a `component`, a normal template (in "templates/") does not work. Which is clear now, after I learned it... But before...
So maybe this helps newcomers to see it on the first glance.